+++ NEWS.md
@@ -15,3 +15,4 @@ version 0.1.0
 =============
 
 * A new package, which takes all primary plotting out of the `SpaDES` package.
+